Added execution config object start true async
The execution config object is added to manage execution
configuration of global kernels.
Basic idea is the following: Given a total number of elements
to process, number of streams and number of threads per block,
we give where each stream's data starts and ends.
From this we compute the number of blocks as well.
The trueasyncgrad.cpp contains some unfinished code to use
execution conf1d object with streams